consolidate pages
