+++ client/scripts/views/pages/subscriptions/notificationSettings.ts
@@ -3,9 +3,12 @@ import 'pages/notification_subscriptions.scss';
 import m from 'mithril';
 import $ from 'jquery';
 import _ from 'lodash';
-import { Button, Icons, ListItem, Table, SelectList, } from 'construct-ui';
+import { Checkbox, Button, Icons, ListItem, Table, SelectList } from 'construct-ui';
+import {
+  SubstrateEvents, SubstrateTypes, IChainEventKind, TitlerFilter
+} from '@commonwealth/chain-events';
 
-import { NotificationSubscription, ChainInfo, CommunityInfo } from 'models';
+import { NotificationSubscription, ChainInfo, CommunityInfo, ChainNetwork } from 'models';
 import app from 'state';
 import { NotificationCategories } from 'types';
 
@@ -13,7 +16,9 @@ import Sublayout from 'views/sublayout';
 import PageLoading from 'views/pages/loading';
 import PageError from 'views/pages/error';
 import { sortSubscriptions } from 'helpers/notifications';
-import { ChainNotificationManagementPage } from './chainEventSettings';
+import {
+  EdgewareChainNotificationTypes, KusamaChainNotificationTypes, PolkdotChainNotificationTypes, KulupuChainNotificationTypes
+} from 'helpers/chain_notification_types';
 
 const NOTIFICATION_TABLE_PRE_COPY = 'Notify me when...';
 const COMMENT_NUM_PREFIX = 'Comment #';
@@ -100,17 +105,13 @@ const batchLabel = (subscriptions: NotificationSubscription[]) => {
     : COMMENT_NUM_PREFIX + threadOrComment.toString();
 };
 
-interface IBatchedSubscriptionRowAttrs {
+const BatchedSubscriptionRow: m.Component<{
   subscriptions: NotificationSubscription[];
   label?: string;
   bold?: boolean;
-}
-
-interface IBatchedSubscriptionRowState {
+}, {
   option: string;
-}
-
-const BatchedSubscriptionRow: m.Component<IBatchedSubscriptionRowAttrs, IBatchedSubscriptionRowState> = {
+}> = {
   view: (vnode) => {
     const { label, bold, subscriptions } = vnode.attrs;
     const someActive = subscriptions.some((s) => s.isActive);
@@ -200,12 +201,271 @@ const NewActivityRow: m.Component = {
   }
 };
 
-interface ICommunitySpecificNotificationsAttrs {
+const EventSubscriptionRow: m.Component<{
+    chain: string;
+    kind: IChainEventKind;
+    titler: TitlerFilter;
+}, {}> = {
+  view: (vnode) => {
+    const { chain, kind } = vnode.attrs;
+    const { title, description } = vnode.attrs.titler(kind);
+    const objectId = `${chain}-${kind}`;
+    const subscription = app.loginStatusLoaded && app.user.notifications.subscriptions
+      .find((sub) => sub.category === NotificationCategories.ChainEvent
+        && sub.objectId === objectId);
+    return m('tr.EventSubscriptionRow', [
+      m('td', `${title}`),
+      app.loginStatusLoaded && m('td', [
+        m(Checkbox, {
+          checked: subscription && subscription.isActive,
+          size: 'lg',
+          onchange: async (e) => {
+            e.preventDefault();
+            if (subscription && subscription.isActive) {
+              await app.user.notifications.disableSubscriptions([ subscription ]);
+            } else if (subscription && !subscription.isActive) {
+              await app.user.notifications.enableSubscriptions([ subscription ]);
+            } else {
+              await app.user.notifications.subscribe(NotificationCategories.ChainEvent, objectId);
+            }
+            m.redraw();
+          }
+        }),
+      ]),
+      m('td', [
+        m(Checkbox, {
+          disabled: !subscription?.isActive,
+          checked: subscription?.isActive && subscription?.immediateEmail,
+          size: 'lg',
+          onchange: async (e) => {
+            e.preventDefault();
+            if (subscription && subscription.immediateEmail) {
+              await app.user.notifications.disableImmediateEmails([ subscription ]);
+            } else {
+              await app.user.notifications.enableImmediateEmails([ subscription ]);
+            }
+            m.redraw();
+          }
+        }),
+      ]),
+    ]);
+  }
+};
+
+/*  This currently is not in production. It is used for testing specific
+    chain events and their related subscriptions. The EventSubscriptionRow
+    above is also only used by IndividualEventSubscriptions.
+*/
+const IndividualEventSubscriptions: m.Component<{
+  chain: ChainNetwork;
+}, {
+  eventKinds: IChainEventKind[];
+  titler;
+  allSupportedChains: string[];
+  isSubscribedAll: boolean;
+  isEmailAll: boolean;
+}> = {
+  oninit: (vnode) => {
+    if (vnode.attrs.chain === ChainNetwork.Edgeware) {
+      vnode.state.titler = SubstrateEvents.Title;
+      vnode.state.eventKinds = SubstrateTypes.EventKinds;
+    } else {
+      vnode.state.titler = null;
+      vnode.state.eventKinds = [];
+    }
+  },
+  view: (vnode) => {
+    const { eventKinds, titler } = vnode.state;
+
+    const supportedChains = app.loginStatusLoaded
+      ? app.config.chains.getAll()
+        .filter((c) => vnode.state.allSupportedChains.includes(c.id))
+        .sort((a, b) => a.id.localeCompare(b.id))
+      : [];
+
+    return [
+      supportedChains.length > 0 && eventKinds.length > 0 && titler
+        ? eventKinds.map((kind) => m(EventSubscriptionRow, {
+          chain: vnode.attrs.chain,
+          kind,
+          titler,
+          key: kind
+        }))
+        : m('No events available on this chain.'),
+    ];
+  },
+};
+
+const EventSubscriptionTypeRow: m.Component<{
+  title: string;
+  notificationTypeArray: string[];
+}, { option: string, }> = {
+  view: (vnode) => {
+    const { title, notificationTypeArray, } = vnode.attrs;
+    const subscriptions = app.user.notifications.subscriptions.filter((s) => {
+      return (
+        s.category === NotificationCategories.ChainEvent
+        && notificationTypeArray.includes(s.objectId)
+      );
+    });
+    const everySubscriptionActive = subscriptions.every((s) => s.isActive);
+    const someSubscriptionsActive = subscriptions.some((s) => s.isActive);
+    const everySubscriptionEmail = subscriptions.every((s) => s.immediateEmail);
+    const someSubscriptionsEmail = subscriptions.some((s) => s.immediateEmail);
+    const allSubscriptionsCreated = subscriptions.length === notificationTypeArray.length;
+
+    if (allSubscriptionsCreated && everySubscriptionActive && everySubscriptionEmail) {
+      vnode.state.option = 'Notifications on (app + email)';
+    } else if (allSubscriptionsCreated && everySubscriptionActive) {
+      vnode.state.option = 'Notifications on (app only)';
+    } else {
+      vnode.state.option = 'Notifications off';
+    }
+
+    return m('tr.EventSubscriptionTypeRow', [
+      m('td', { class: 'bold' }, title),
+      m('td', [
+        m(SelectList, {
+          class: 'EventSubscriptionTypeSelectList',
+          filterable: false,
+          checkmark: false,
+          emptyContent: null,
+          inputAttrs: {
+            class: 'EventSubscriptionTypeSelectRow',
+          },
+          itemRender: (option: string) => {
+            return m(ListItem, {
+              label: option,
+              selected: (vnode.state.option === option),
+            });
+          },
+          items: ['Notifications off', 'Notifications on (app only)', 'Notifications on (app + email)', ],
+          trigger: m(Button, {
+            align: 'left',
+            compact: true,
+            iconRight: Icons.CHEVRON_DOWN,
+            label: vnode.state.option,
+          }),
+          onSelect: async (option: string) => {
+            vnode.state.option = option;
+            if (option === 'Notifications off') {
+              await app.user.notifications.disableImmediateEmails(subscriptions);
+              await app.user.notifications.disableSubscriptions(subscriptions);
+            } else if (option === 'Notifications on (app only)') {
+              if (!allSubscriptionsCreated) {
+                await Promise.all(
+                  notificationTypeArray.map((obj) => {
+                    return app.user.notifications.subscribe(NotificationCategories.ChainEvent, obj);
+                  })
+                );
+              } else {
+                if (!everySubscriptionActive) await app.user.notifications.enableSubscriptions(subscriptions);
+              }
+              if (someSubscriptionsEmail) await app.user.notifications.disableImmediateEmails(subscriptions);
+            } else if (option === 'Notifications on (app + email)') {
+              if (!allSubscriptionsCreated) {
+                await Promise.all(
+                  notificationTypeArray.map((obj) => {
+                    return app.user.notifications.subscribe(NotificationCategories.ChainEvent, obj);
+                  })
+                ).then(async () => {
+                  const newSubscriptions = app.user.notifications.subscriptions.filter((s) => {
+                    return (
+                      s.category === NotificationCategories.ChainEvent
+                      && notificationTypeArray.includes(s.objectId)
+                    );
+                  });
+                  await app.user.notifications.enableImmediateEmails(newSubscriptions);
+                  m.redraw();
+                });
+              } else {
+                if (!everySubscriptionActive) await app.user.notifications.enableSubscriptions(subscriptions);
+                if (!everySubscriptionEmail) await app.user.notifications.enableImmediateEmails(subscriptions);
+              }
+            }
+            m.redraw();
+          }
+        }),
+      ]),
+    ]);
+  }
+};
+
+const EdgewareChainEvents: m.Component = {
+  view: (vnode) => {
+    return [
+      m(EventSubscriptionTypeRow, { title: 'Council events', notificationTypeArray: EdgewareChainNotificationTypes.Council, }),
+      m(EventSubscriptionTypeRow, { title: 'Democracy events', notificationTypeArray: EdgewareChainNotificationTypes.Democracy, }),
+      m(EventSubscriptionTypeRow, { title: 'Preimage events', notificationTypeArray: EdgewareChainNotificationTypes.Preimage, }),
+      m(EventSubscriptionTypeRow, { title: 'Signaling events', notificationTypeArray: EdgewareChainNotificationTypes.Signaling, }),
+      m(EventSubscriptionTypeRow, { title: 'Treasury events', notificationTypeArray: EdgewareChainNotificationTypes.Treasury, }),
+      m(EventSubscriptionTypeRow, { title: 'Validator events', notificationTypeArray: EdgewareChainNotificationTypes.Validator, }),
+      m(EventSubscriptionTypeRow, { title: 'Vote events', notificationTypeArray: EdgewareChainNotificationTypes.Vote, }),
+    ];
+  }
+};
+
+const KusamaChainEvents: m.Component = {
+  view: (vnode) => {
+    return [
+      m(EventSubscriptionTypeRow, { title: 'Council events', notificationTypeArray: KusamaChainNotificationTypes.Council, }),
+      m(EventSubscriptionTypeRow, { title: 'Democracy events', notificationTypeArray: KusamaChainNotificationTypes.Democracy, }),
+      m(EventSubscriptionTypeRow, { title: 'Preimage events', notificationTypeArray: KusamaChainNotificationTypes.Preimage, }),
+      // m(EventSubscriptionTypeRow, { title: 'Treasury events', notificationTypeArray: KusamaChainNotificationTypes.Treasury, }),
+      m(EventSubscriptionTypeRow, { title: 'Validator events', notificationTypeArray: KusamaChainNotificationTypes.Validator, }),
+      m(EventSubscriptionTypeRow, { title: 'Vote events', notificationTypeArray: KusamaChainNotificationTypes.Vote, }),
+    ];
+  }
+};
+
+const PolkadotChainEvents: m.Component = {
+  view: (vnode) => {
+    return [
+      m(EventSubscriptionTypeRow, { title: 'Council events', notificationTypeArray: PolkdotChainNotificationTypes.Council, }),
+      m(EventSubscriptionTypeRow, { title: 'Democracy events', notificationTypeArray: PolkdotChainNotificationTypes.Democracy, }),
+      m(EventSubscriptionTypeRow, { title: 'Preimage events', notificationTypeArray: PolkdotChainNotificationTypes.Preimage, }),
+      // m(EventSubscriptionTypeRow, { title: 'Treasury events', notificationTypeArray: PolkdotChainNotificationTypes.Treasury, }),
+      m(EventSubscriptionTypeRow, { title: 'Validator events', notificationTypeArray: PolkdotChainNotificationTypes.Validator, }),
+      m(EventSubscriptionTypeRow, { title: 'Vote events', notificationTypeArray: PolkdotChainNotificationTypes.Vote, }),
+    ];
+  }
+};
+
+const KulupuChainEvents: m.Component = {
+  view: (vnode) => {
+    return [
+      m(EventSubscriptionTypeRow, { title: 'Council events', notificationTypeArray: KulupuChainNotificationTypes.Council, }),
+      m(EventSubscriptionTypeRow, { title: 'Democracy events', notificationTypeArray: KulupuChainNotificationTypes.Democracy, }),
+      m(EventSubscriptionTypeRow, { title: 'Preimage events', notificationTypeArray: KulupuChainNotificationTypes.Preimage, }),
+      // m(EventSubscriptionTypeRow, { title: 'Treasury events', notificationTypeArray: PolkdotChainNotificationTypes.Treasury, }),
+      m(EventSubscriptionTypeRow, { title: 'Validator events', notificationTypeArray: KulupuChainNotificationTypes.Validator, }),
+      m(EventSubscriptionTypeRow, { title: 'Vote events', notificationTypeArray: KulupuChainNotificationTypes.Vote, }),
+    ];
+  }
+};
+
+const EventSubscriptions: m.Component<{chain: ChainInfo}> = {
+  view: (vnode) => {
+    const { chain } = vnode.attrs;
+    return m('.EventSubscriptions', [
+      m(Table, {}, [
+        m('tr', [
+          m('th', null),
+          m('th', 'Settings'),
+        ]),
+        (chain.network === ChainNetwork.Edgeware) && m(EdgewareChainEvents),
+        (chain.network === ChainNetwork.Kulupu) && m(KulupuChainEvents),
+        (chain.network === ChainNetwork.Kusama) && m(KusamaChainEvents),
+        (chain.network === ChainNetwork.Polkadot) && m(PolkadotChainEvents),
+      ]),
+    ]);
+  }
+};
+
+const CommunitySpecificNotifications: m.Component<{
   community: CommunityInfo | ChainInfo;
   subscriptions: NotificationSubscription[];
-}
-
-const CommunitySpecificNotifications: m.Component<ICommunitySpecificNotificationsAttrs, {}> = {
+}, {}> = {
   view: (vnode) => {
     const { community, subscriptions } = vnode.attrs;
     const filteredSubscriptions = subscriptions.filter(
@@ -233,43 +493,37 @@ const CommunitySpecificNotifications: m.Component<ICommunitySpecificNotification
   },
 };
 
-interface IGeneralNewThreadsAndCommentsAttrs {
+const GeneralNewThreadsAndComments:
+  m.Component<{
   communities: CommunityInfo[];
   subscriptions: NotificationSubscription[];
-}
-
-interface IGeneralCommunityNotificationsState {
+}, {
   generalStatus: boolean;
   emailStatus: boolean;
   generalOpen: boolean;
   emailOpen: boolean;
-}
+}> = {
+  oninit: (vnode) => {
+    vnode.state.generalStatus = null;
+    vnode.state.emailStatus = null;
+    vnode.state.generalOpen = false;
+    vnode.state.emailStatus = false;
+  },
+  view: (vnode) => {
+    const { communities, subscriptions } = vnode.attrs;
+    const communityIds = communities.map((c) => c.id);
+    const threadSubs = subscriptions.filter((s) => communityIds.includes(s.objectId));
+    return m(BatchedSubscriptionRow, {
+      subscriptions: threadSubs,
+      label: NEW_THREADS_ALL_COMMUNITIES_LABEL,
+    });
+  },
+};
 
-const GeneralNewThreadsAndComments:
-  m.Component<IGeneralNewThreadsAndCommentsAttrs, IGeneralCommunityNotificationsState > = {
-    oninit: (vnode) => {
-      vnode.state.generalStatus = null;
-      vnode.state.emailStatus = null;
-      vnode.state.generalOpen = false;
-      vnode.state.emailStatus = false;
-    },
-    view: (vnode) => {
-      const { communities, subscriptions } = vnode.attrs;
-      const communityIds = communities.map((c) => c.id);
-      const threadSubs = subscriptions.filter((s) => communityIds.includes(s.objectId));
-      return m(BatchedSubscriptionRow, {
-        subscriptions: threadSubs,
-        label: NEW_THREADS_ALL_COMMUNITIES_LABEL,
-      });
-    },
-  };
-
-interface IGeneralCommunityNotificationsAttrs {
+const GeneralCommunityNotifications: m.Component<{
   subscriptions: NotificationSubscription[];
   communities: CommunityInfo[];
-}
-
-const GeneralCommunityNotifications: m.Component<IGeneralCommunityNotificationsAttrs> = {
+}> = {
   view: (vnode) => {
     const { subscriptions, communities } = vnode.attrs;
     const mentionsSubscription = subscriptions.find((s) => s.category === NotificationCategories.NewMention);
@@ -290,24 +544,19 @@ const GeneralCommunityNotifications: m.Component<IGeneralCommunityNotificationsA
       batchedSubscriptions.map((subscriptions2: NotificationSubscription[]) => {
         return m(BatchedSubscriptionRow, { subscriptions: subscriptions2 });
       })
-      // m(GeneralPastThreadsAndComments, { subscriptions }),
     ];
   },
 };
 
-interface ICommunityNotificationsAttrs {
+const CommunityNotifications: m.Component<{
   subscriptions: NotificationSubscription[];
   communities: CommunityInfo[];
   chains: ChainInfo[];
-}
-
-interface ICommunityNotificationsState {
+}, {
   selectedCommunity: CommunityInfo | ChainInfo;
   selectedCommunityId: string;
   communityIds: string[];
-}
-
-const CommunityNotifications: m.Component<ICommunityNotificationsAttrs, ICommunityNotificationsState> = {
+}> = {
   oninit: (vnode) => {
     vnode.state.communityIds = [ALL_COMMUNITIES];
     vnode.attrs.communities.forEach((c) => vnode.state.communityIds.push(c.name));
@@ -371,6 +620,54 @@ const CommunityNotifications: m.Component<ICommunityNotificationsAttrs, ICommuni
       ]),
     ]);
   }
+};
+
+export const ChainNotificationManagementPage: m.Component<{ chains: ChainInfo[] }, { selectedChain: ChainInfo }> = {
+  oninit: (vnode) => {
+    const { chains } = vnode.attrs;
+    const scope = m.route.param('scope');
+    vnode.state.selectedChain = chains.find((c) => c.id === scope)
+      || chains.find((c) => c.id === 'edgeware');
+  },
+  view: (vnode) => {
+    const { chains } = vnode.attrs;
+    if (chains.length < 1) return;
+    const validChains = [ChainNetwork.Edgeware, ChainNetwork.Polkadot, ChainNetwork.Kusama, ChainNetwork.Kulupu];
+    const filteredChains = chains.filter((c) => validChains.includes(c.network)).sort((a, b) => (a.id > b.id) ? 1 : -1);
+    return m('ChainNotificationManagementPage', [
+      m(SelectList, {
+        class: 'ChainNotificationSelectList',
+        filterable: false,
+        checkmark: false,
+        emptyContent: null,
+        inputAttrs: {
+          class: 'CommunitySelectRow',
+        },
+        itemRender: (chain: ChainInfo) => {
+          return m(ListItem, {
+            label: chain.name,
+            selected: (vnode.state.selectedChain === chain),
+          });
+        },
+        items: filteredChains,
+        trigger: m(Button, {
+          align: 'left',
+          compact: true,
+          iconRight: Icons.CHEVRON_DOWN,
+          label: vnode.state.selectedChain
+            ? vnode.state.selectedChain.name
+            : 'Select Chain',
+        }),
+        onSelect: (chain: ChainInfo) => {
+          vnode.state.selectedChain = chain;
+          m.redraw();
+        }
+      }),
+      m(EventSubscriptions, {
+        chain: vnode.state.selectedChain,
+      }),
+    ]);
+  },
 };
 
 const NotificationSettingsPage: m.Component<{}, {
