--- client/styles/pages/subscriptions.scss
@@ -1,65 +1,0 @@
-@import 'client/styles/shared';
-
-.SubscriptionsPage {
-    .user-feedback {
-        margin-left: 12px;
-        color: #444;
-    }
-
-    .SubscriptionRow {
-        border: solid 1px $background-color-light;
-        border-radius: 3px;
-        margin: 40px 0;
-        padding: 10px 20px 20px;
-        box-shadow: 0px 2px 4px -2px rgba(67, 90, 111, 0.3);
-
-        td:nth-child(1) {
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-
-            a {
-                white-space: nowrap;
-            }
-        }
-
-        h4 {
-            margin-top: 0;
-            margin-bottom: 8px;
-        }
-
-        button {
-            margin: 0 8px 0 0;
-        }
-    }
-
-    .NotificationButtons {
-        button {
-            margin: 8px 8px 0 0;
-        }
-    }
-
-    .PauseButtons {
-        .PauseToggle {
-            margin: 8px 8px 0 0;
-        }
-    }
-
-    tr {
-        th {
-            text-align: center;
-        }
-        td {
-            a, a:visited {
-                color: black;
-            }
-        }
-        td:nth-child(2), td:nth-child(3)  {
-            text-align: center;
-            label {
-                margin: 0;
-                padding: 9px;
-            }
-        }
-    }
-}
